+++ src/lucuma/core.cljs
@@ -197,9 +197,7 @@
   ; Set default properties values
   (install-lucuma-properties-holder! el)
   (install-properties-holder! el)
-  (set-properties! el m ps true true)
-  (doseq [[k v] (:attributes ps)]
-    (.setAttribute el (name k) (str v))))
+  (set-properties! el m ps true true))
 
 (defn- install-content!
   [el {:keys [document style]}]
@@ -315,7 +313,7 @@
         (merge m {:type it})))))
 
 (def all-keys
-  #{:name :ns :prototype :extends :mixins :document :style :attributes :properties :methods
+  #{:name :ns :prototype :extends :mixins :document :style :properties :methods
     :on-created :on-attached :on-detached :on-property-changed})
 
 (defn ignored-keys