+++ test/lucuma/core_test.cljs
@@ -229,12 +229,6 @@
   (is (= "rgb(255, 0, 0)" (.-color (.getComputedStyle js/window (by-id "style1")))))
   (is (= "rgb(255, 0, 0)" (.-color (.getComputedStyle js/window (by-id "style2"))))))
 
-(defwebcomponent test-attribute-1
-  :attributes {:attribute 1})
-
-(deftest attributes
-  (is (= "1" (.getAttribute (.createElement js/document "test-attribute-1") "attribute"))))
-
 (defwebcomponent test-on-created-1
   :on-created (fn [] {:document "content"}))
 (defwebcomponent test-on-created-2
@@ -482,8 +476,6 @@
   (l/register test-style-2)
   (append "test-style-1")
   (append "test-style-2")
-
-  (l/register test-attribute-1)
 
   (l/register test-on-created-1)
   (l/register test-on-created-2)